# models/create_device_request.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateDeviceRequest CreateDeviceRequest
//
// swagger:model CreateDeviceRequest
type CreateDeviceRequest struct {

	// The brand of the device

	Brand string `json:"brand,omitempty"`

	// The name of the device

	Name string `json:"name,omitempty"`
}

// Validate validates this create device request
func (m *CreateDeviceRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create device request based on context it is used
func (m *CreateDeviceRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateDeviceRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateDeviceRequest) UnmarshalBinary(b []byte) error {
	var res CreateDeviceRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}