# models/create_device_response.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateDeviceResponse CreateDeviceResponse
//
// swagger:model CreateDeviceResponse
type CreateDeviceResponse struct {

	// The id of the device

	ID string `json:"id,omitempty"`

	// The name of the device

	Name string `json:"name,omitempty"`
}

// Validate validates this create device response
func (m *CreateDeviceResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create device response based on context it is used
func (m *CreateDeviceResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateDeviceResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateDeviceResponse) UnmarshalBinary(b []byte) error {
	var res CreateDeviceResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}