# models/device_brand_update_request.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// DeviceBrandUpdateRequest DeviceBrandUpdateRequest
//
// swagger:model DeviceBrandUpdateRequest
type DeviceBrandUpdateRequest struct {

	// The brand of the device

	Brand string `json:"brand,omitempty"`
}

// Validate validates this device brand update request
func (m *DeviceBrandUpdateRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this device brand update request based on context it is used
func (m *DeviceBrandUpdateRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *DeviceBrandUpdateRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *DeviceBrandUpdateRequest) UnmarshalBinary(b []byte) error {
	var res DeviceBrandUpdateRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}