# models/update_device_request.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// UpdateDeviceRequest UpdateDeviceRequest
//
// swagger:model UpdateDeviceRequest
type UpdateDeviceRequest struct {

	// The brand of the device

	Brand string `json:"brand,omitempty"`

	// The name of the device

	Name string `json:"name,omitempty"`
}

// Validate validates this update device request
func (m *UpdateDeviceRequest) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this update device request based on context it is used
func (m *UpdateDeviceRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *UpdateDeviceRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdateDeviceRequest) UnmarshalBinary(b []byte) error {
	var res UpdateDeviceRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}