# swagger.yml
basePath: /device
definitions:
  Device:
    properties:
      id:
        description: The id of the device
        type: string
        x-go-name: ID
      name:
        description: The name of the device
        type: string
        x-go-name: Name
      brand:
        description: The brand of the device
        type: string
        x-go-name: Brand
      createdAt:
        description: The time the device was created
        type: string
        format: date-time
        x-go-name: CreatedAt
    title: Device
    type: object
  CreateDeviceRequest:
    properties:
      name:
        description: The name of the device
        type: string
        x-go-name: Name
      brand:
        description: The brand of the device
        type: string
        x-go-name: Brand
    title: CreateDeviceRequest
    type: object
  CreateDeviceResponse:
    properties:
      id:
        description: The id of the device
        type: string
        x-go-name: ID
      name:
        description: The name of the device
        type: string
        x-go-name: Name
    title: CreateDeviceResponse
    type: object
  UpdateDeviceRequest:
    properties:
      name:
        description: The name of the device
        type: string
        x-go-name: Name
      brand:
        description: The brand of the device
        type: string
        x-go-name: Brand
    title: UpdateDeviceRequest
  DeviceBrandUpdateRequest:
    properties:
      brand:
        description: The brand of the device
        type: string
        x-go-name: Brand
    title: DeviceBrandUpdateRequest
  DeviceNameUpdateRequest:
    properties:
      name:
        description: The name of the device
        type: string
        x-go-name: Name
    title: DeviceNameUpdateRequest
  Error:
    description: An error in a request
    properties:
      code:
        type: number
      message:
        type: string
      result:
        type: boolean
    type: object
info:
  description: |
    The purpose of this microservice is to keep the record of devices.
    | Error                                   | Code |
    | --------------------------------------- | -    |
    | requiredParameter                       | 1    |
    | invalidParameter                        | 2    |
    | createError                             | 3    |
    | listError                               | 4    |
    | couldNotFindObject                      | 5    |
    | updateError                             | 6    |
    | deleteError                             | 7    |
    | decodeError                             | 8    |
  title: device
  version: v1
paths:
  /:
    get:
      consumes:
        - application/json
      description: this endpoint returns the devices
      operationId: getDevices
      produces:
        - application/json
      parameters:
        - description: The brand of the device
          in: query
          name: brand
          required: false
          type: string
      responses:
        "200":
          description: success response
          schema:
            items:
              $ref: "#/definitions/Device"
            type: array
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      tags:
        - Device
    post:
      consumes:
        - application/json
      description: this endpoint creates a device
      operationId: createDevice
      parameters:
        - in: body
          name: device creation request body
          required: true
          schema:
            $ref: "#/definitions/CreateDeviceRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created device id and name
          schema:
            $ref: "#/definitions/CreateDeviceResponse"
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      security:
        - venueId: []
      tags:
        - Device
  /{id}:
    get:
      consumes:
        - application/json
      description: this endpoint returns a device for a given id
      operationId: getDevice
      parameters:
        - description: device id or number
          name: id
          in: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: success response
          schema:
            $ref: "#/definitions/Device"
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      tags:
        - Device
    put:
      consumes:
        - application/json
      description: update device
      operationId: updateDevice
      parameters:
        - description: device's ID
          in: path
          name: id
          required: true
          type: string
        - in: body
          name: device update request body
          required: true
          schema:
            $ref: "#/definitions/UpdateDeviceRequest"
      produces:
        - application/json
      responses:
        "204":
          description: success no content
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      security:
        - profileId: []
      tags:
        - Device
    delete:
      consumes:
        - application/json
      description: delete device
      operationId: deleteDevice
      parameters:
        - description: device's ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "204":
          description: success no content
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      security:
        - profileId: []
      tags:
        - Device
  /{id}/name:
    put:
      consumes:
        - application/json
      description: this endpoint adjusts the name of a device
      operationId: updateDeviceName
      parameters:
        - description: The id of the device to update the name for
          in: path
          name: id
          required: true
          type: string
        - in: body
          description: Device name update request
          name: deviceNameUpdate
          schema:
            $ref: "#/definitions/DeviceNameUpdateRequest"
      produces:
        - application/json
      responses:
        "204":
          description: success no content
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      tags:
        - Device
  /{id}/brand:
    put:
      consumes:
        - application/json
      description: this endpoint adjusts the brand of a device
      operationId: updateDeviceBrand
      parameters:
        - description: The id of the device to update the brand for
          in: path
          name: id
          required: true
          type: string
        - in: body
          description: Device brand update request
          name: deviceBrandUpdate
          schema:
            $ref: "#/definitions/DeviceBrandUpdateRequest"
      produces:
        - application/json
      responses:
        "204":
          description: success no content
        "400":
          description: Required parameters were not sent
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: Object does not exist
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: A problem when processing the request
          schema:
            $ref: "#/definitions/Error"
      tags:
        - Device
schemes:
  - https
swagger: "2.0"